Städat samt buggfixat
Jag har städat upp en del kod där jag brutit ut lite saker från bodyn på de olika filerna i screens. Har även fixat lite designaspekter i focus_recipe_view.
